fix/filebeat-certs
* retrieves filebeat CA from indexer relation.

the charm originally assumed the same CA would be used for the rsyslog receiver
certificate and for the Wazuh indexer; this is not necessarily true,
particularly if a non-self-signed cert is being used for the public-facing
rsyslog receiver.

* adds an integration test for filebeat output.

* refactors unit tests for state.

new tests are much less repetitive, differentiate get_secret results based on
the id/label provided (not true of the original), and generally test the state
more thoroughly with fewer LOCs.