fix/filebeat-certs (#324)
* fix/filebeat-certs

* retrieves filebeat CA from indexer relation.

the charm originally assumed the same CA would be used for the rsyslog receiver
certificate and for the Wazuh indexer; this is not necessarily true,
particularly if a non-self-signed cert is being used for the public-facing
rsyslog receiver.

* adds an integration test for filebeat output.

* refactors unit tests for state.

new tests are much less repetitive, differentiate get_secret results based on
the id/label provided (not true of the original), and generally test the state
more thoroughly with fewer LOCs.

* Ignore CVEs not under our control

* update integration test model refs for certs

* fix/filebeat-certs test fixes

* fix/filebeat-certs test fixes pt2

* accept suggestion: take last cert in chain

---------

Co-authored-by: Sebastien Georget <sebastien.georget@canonical.com>